// packages/neutral/src/generated/accounts/bundle.ts
export const BUNDLE_DISCRIMINATOR = new Uint8Array([
  15, 82, 167, 230, 37, 214, 82, 80,
]);

export function getBundleDiscriminatorBytes() {
  return fixEncoderSize(getBytesEncoder(), 8).encode(BUNDLE_DISCRIMINATOR);
}

export type Bundle = {
  discriminator: ReadonlyUint8Array;
  name: ReadonlyUint8Array;
  manager: Address;
  keeper: Address;
  treasuryAccount: Address;
  allocatedReceivers: Array<Address>;
  bundleUnderlyingBalance: bigint;
  maxDepositAmount: bigint;
  withdrawalDelay: bigint;
  performanceFee: number;
  managementFeeBps: number;
  depositFee: number;
  withdrawalFee: number;
  managerPfeeShares: bigint;
  currentAllocationBps: number;
  oracleBuffer: bigint;
  totalShares: bigint;
  assetPrecision: bigint;
  assetAddress: Address;
  assetDecimals: number;
  withdrawalTMin: bigint;
  withdrawalTMax: bigint;
  withdrawalCurve: number;
  permissionned: boolean;
  managerMfeeShares: bigint;
  minDepositAmount: bigint;
  oracleUpdateTimeLimit: bigint;
  oracleMaxAge: bigint;
  withdrawalRedemptionRequestCutoffTs: bigint;
  withdrawalRedemptionUnlockCurrentCycleTs: bigint;
  withdrawalRedemptionUnlockNextCycleTs: bigint;
  referrerEnabled: boolean;
  referrerMinDepositAmount: bigint;
  referralTiers: Array<ReferralTier>;
  tierCount: number;
  padding: ReadonlyUint8Array;
};