// packages/neutral/src/generated/accounts/userBundleAccount.ts
export const USER_BUNDLE_ACCOUNT_DISCRIMINATOR = new Uint8Array([
  32, 181, 106, 26, 67, 130, 185, 241,
]);

export function getUserBundleAccountDiscriminatorBytes() {
  return fixEncoderSize(getBytesEncoder(), 8).encode(
    USER_BUNDLE_ACCOUNT_DISCRIMINATOR,
  );
}

export type UserBundleAccount = {
  discriminator: ReadonlyUint8Array;
  owner: Address;
  lastDepositTimestamp: bigint;
  shares: bigint;
  pendingDeposit: bigint;
  pendingShares: bigint;
  estimatedPendingWithdrawalValue: bigint;
  withdrawalAvailableTimestamp: bigint;
  lastWithdrawalProcessTimestamp: bigint;
  lastHighWaterMark: bigint;
  hwmPerShare: bigint;
  lastManagementFeeTimestamp: bigint;
  netDeposits: bigint;
  totalFeeCharged: bigint;
  customDepositFeeBps: number;
  customWithdrawalFeeBps: number;
  customPerformanceFeeBps: number;
  customManagementFeeBps: number;
  feeOverrideFlags: number;
  customWithdrawalDelay: bigint;
  customWithdrawalTMin: bigint;
  customWithdrawalTMax: bigint;
  customWithdrawalCurve: number;
  withdrawalTimingOverrideFlags: number;
  switchActive: boolean;
  switchTargetBundle: Address;
  switchCreatedAt: bigint;
  referrer: Address;
  padding: ReadonlyUint8Array;
};